refactor: Refactor delete category use case and repository

+++ src/application/use-cases/category/delete-category.use-case.ts
@@ -1,9 +1,16 @@
 import { ICategoryRepository } from "@/application/repositories/category.repository.interface";
+import { DBNotFoundError } from "@/entities/errors/database.error";
 
 export type IDeleteCategoryUseCase = ReturnType<typeof deleteCategoryUseCase>;
 
 export const deleteCategoryUseCase =
   (categoryRepository: ICategoryRepository) =>
   async (categoryId: number): Promise<void> => {
+    const category = await categoryRepository.getById(categoryId);
+
+    if (!category) {
+      throw new DBNotFoundError("Category not found");
+    }
+
     return await categoryRepository.delete(categoryId);
   };
